requirements.txt
fastapi
uvicorn
python-dotenv
weaviate-client
sentence-transformers
transformers
pydantic
unidecode
python-frontmatter
markdown
requests
